refactor upload service constants and types
