+++ src/constants/upload/index.ts
@@ -1,5 +1,6 @@
 import { ENCRYPTION_CHUNK_SIZE } from 'constants/crypto';
 import { FILE_TYPE } from 'constants/file';
+import { Location } from 'types/upload';
 
 // list of format that were missed by type-detection for some files.
 export const FORMAT_MISSED_BY_FILE_TYPE_LIB = [
@@ -18,3 +19,21 @@ export const FILE_CHUNKS_COMBINED_FOR_A_UPLOAD_PART = Math.floor(
 );
 
 export const RANDOM_PERCENTAGE_PROGRESS_FOR_PUT = () => 90 + 10 * Math.random();
+
+export const NULL_LOCATION: Location = { latitude: null, longitude: null };
+
+export enum UPLOAD_STAGES {
+    START,
+    READING_GOOGLE_METADATA_FILES,
+    UPLOADING,
+    FINISH,
+}
+
+export enum FileUploadResults {
+    FAILED = -1,
+    SKIPPED = -2,
+    UNSUPPORTED = -3,
+    BLOCKED = -4,
+    TOO_LARGE = -5,
+    UPLOADED = 100,
+}
