+++ src/services/upload/uiService.ts
@@ -1,18 +1,8 @@
-import { RANDOM_PERCENTAGE_PROGRESS_FOR_PUT } from 'constants/upload';
-import { UPLOAD_STAGES } from './uploadManager';
-
-export interface ProgressUpdater {
-    setPercentComplete: React.Dispatch<React.SetStateAction<number>>;
-    setFileCounter: React.Dispatch<
-        React.SetStateAction<{
-            finished: number;
-            total: number;
-        }>
-    >;
-    setUploadStage: React.Dispatch<React.SetStateAction<UPLOAD_STAGES>>;
-    setFileProgress: React.Dispatch<React.SetStateAction<Map<string, number>>>;
-    setUploadResult: React.Dispatch<React.SetStateAction<Map<string, number>>>;
-}
+import {
+    RANDOM_PERCENTAGE_PROGRESS_FOR_PUT,
+    UPLOAD_STAGES,
+} from 'constants/upload';
+import { ProgressUpdater } from 'types/upload';
 
 class UIService {
     private perFileProgress: number;
