+++ src/services/upload/uploadService.ts
@@ -1,69 +1,29 @@
-import { fileAttribute } from 'types/file';
 import { Collection } from 'types/collection';
 import { logError } from 'utils/sentry';
 import UploadHttpClient from './uploadHttpClient';
-import {
-    extractMetadata,
-    getMetadataMapKey,
-    ParsedMetaDataJSON,
-} from './metadataService';
+import { extractMetadata, getMetadataMapKey } from './metadataService';
 import { generateThumbnail } from './thumbnailService';
-import {
-    getFileOriginalName,
-    getFileData,
-    FileTypeInfo,
-} from './readFileService';
+import { getFileOriginalName, getFileData } from './readFileService';
 import { encryptFiledata } from './encryptionService';
 import { uploadStreamUsingMultipart } from './multiPartUploadService';
 import UIService from './uiService';
 import { handleUploadError } from 'utils/error';
-import { MetadataMap } from './uploadManager';
 import {
-    DataStream,
+    B64EncryptionResult,
+    BackupedFile,
+    EncryptedFile,
     EncryptionResult,
+    FileInMemory,
+    FileTypeInfo,
+    FileWithMetadata,
     isDataStream,
+    MetadataMap,
     MetadataObject,
+    ParsedMetaDataJSON,
+    ProcessedFile,
+    UploadFile,
+    UploadURL,
 } from 'types/upload';
-
-export interface UploadURL {
-    url: string;
-    objectKey: string;
-}
-
-export interface B64EncryptionResult {
-    encryptedData: string;
-    key: string;
-    nonce: string;
-}
-
-export interface FileInMemory {
-    filedata: Uint8Array | DataStream;
-    thumbnail: Uint8Array;
-    hasStaticThumbnail: boolean;
-}
-
-export interface FileWithMetadata
-    extends Omit<FileInMemory, 'hasStaticThumbnail'> {
-    metadata: MetadataObject;
-}
-
-export interface EncryptedFile {
-    file: ProcessedFile;
-    fileKey: B64EncryptionResult;
-}
-export interface ProcessedFile {
-    file: fileAttribute;
-    thumbnail: fileAttribute;
-    metadata: fileAttribute;
-    filename: string;
-}
-export interface BackupedFile extends Omit<ProcessedFile, 'filename'> {}
-
-export interface UploadFile extends BackupedFile {
-    collectionID: number;
-    encryptedKey: string;
-    keyDecryptionNonce: string;
-}
 
 class UploadService {
     private uploadURLs: UploadURL[] = [];
