+++ src/types/upload/index.ts
@@ -1,4 +1,6 @@
 import { FILE_TYPE } from 'constants/file';
+import { UPLOAD_STAGES } from 'constants/upload';
+import { Collection } from 'types/collection';
 import { fileAttribute } from 'types/file';
 
 export interface DataStream {
@@ -23,4 +25,88 @@ export interface MetadataObject {
     longitude: number;
     fileType: FILE_TYPE;
     hasStaticThumbnail?: boolean;
+}
+
+export interface Location {
+    latitude: number;
+    longitude: number;
+}
+
+export interface ParsedMetaDataJSON {
+    creationTime: number;
+    modificationTime: number;
+    latitude: number;
+    longitude: number;
+}
+
+export interface MultipartUploadURLs {
+    objectKey: string;
+    partURLs: string[];
+    completeURL: string;
+}
+
+export interface FileTypeInfo {
+    fileType: FILE_TYPE;
+    exactType: string;
+}
+
+export interface ProgressUpdater {
+    setPercentComplete: React.Dispatch<React.SetStateAction<number>>;
+    setFileCounter: React.Dispatch<
+        React.SetStateAction<{
+            finished: number;
+            total: number;
+        }>
+    >;
+    setUploadStage: React.Dispatch<React.SetStateAction<UPLOAD_STAGES>>;
+    setFileProgress: React.Dispatch<React.SetStateAction<Map<string, number>>>;
+    setUploadResult: React.Dispatch<React.SetStateAction<Map<string, number>>>;
+}
+
+export interface FileWithCollection {
+    file: File;
+    collectionID?: number;
+    collection?: Collection;
+}
+
+export type MetadataMap = Map<string, ParsedMetaDataJSON>;
+
+export interface UploadURL {
+    url: string;
+    objectKey: string;
+}
+
+export interface B64EncryptionResult {
+    encryptedData: string;
+    key: string;
+    nonce: string;
+}
+
+export interface FileInMemory {
+    filedata: Uint8Array | DataStream;
+    thumbnail: Uint8Array;
+    hasStaticThumbnail: boolean;
+}
+
+export interface FileWithMetadata
+    extends Omit<FileInMemory, 'hasStaticThumbnail'> {
+    metadata: MetadataObject;
+}
+
+export interface EncryptedFile {
+    file: ProcessedFile;
+    fileKey: B64EncryptionResult;
+}
+export interface ProcessedFile {
+    file: fileAttribute;
+    thumbnail: fileAttribute;
+    metadata: fileAttribute;
+    filename: string;
+}
+export interface BackupedFile extends Omit<ProcessedFile, 'filename'> {}
+
+export interface UploadFile extends BackupedFile {
+    collectionID: number;
+    encryptedKey: string;
+    keyDecryptionNonce: string;
 }
